archive: list: fix segfault when printing table
The segfault is caused by crate `prettytable-rs`,
but in the former versions of `archive`, it works
fine. I tried checking out the older version,
but it still crashes (I want to test it with
an older `prettytable-rs` version tracked in
Cargo.lock).

The `prettytable-rs` crate is at version 0.9.0,
and I tried to upgrade it to 0.10.0, the segfault
is gone. Anyway this problem was fixed by upgrading
its
version.